Migrate from MathJax to Katex
I had a `-` instead of a `_` when setting up the math engine for
kramdown! I prefer this strategy, shipping the minified Katex CSS and
fonts, so that the site does not depend on a CDN and the math equations
are rendered up front (Instead of on every client, as there is no need
to do the same work repeatedly).

I snuck in some minor fixes to the previous posts, but no significant
changes. I should look into adding a "last edited" field to the posts.